Create VideoList.tsx
